feat: create useAudioPreviewManager
